# docs/open/openapi.yml
openapi: "3.0.3"
info:
  title: "HTTP API"
  version: "6.0"
servers:
  - url: "http://localhost:5000/api/v6"
security:
  - jwt: []
  - ApiKeyAuth: []
...
paths:
...
  /rate:
    get:
      tags:
        - rate
      summary: List rates
      parameters:
        - name: rate_type
          in: query
          schema:
            type: string
            enum: ["RATE_RATE", "RATE_BUY", "RATE_SELL", "RATE_AVERAGE"]
        - name: currency_code
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - rate
      summary: Create a new rate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/RatePost"
      responses:
        "201":
          description: Rate created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rate"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /tax:
    get:
      tags:
        - tax
      summary: List taxes
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - tax
      summary: Create a new tax
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/TaxPost"
      responses:
        "201":
          description: Tax created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tax"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /tool:
    get:
      tags:
        - tool
      summary: List tools
      parameters:
        - name: description
          in: query
          schema:
            type: string
        - name: product_code
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - tool
      summary: Create a new tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/ToolPost"
      responses:
        "201":
          description: Tool created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /trans:
    get:
      tags:
        - trans
      summary: List transactions
      parameters:
        - name: trans_type
          in: query
          schema:
            type: string
            enum: ["TRANS_INVOICE", "TRANS_RECEIPT", "TRANS_ORDER", "TRANS_OFFER", "TRANS_WORKSHEET", "TRANS_RENT", "TRANS_DELIVERY", "TRANS_INVENTORY", "TRANS_WAYBILL", "TRANS_PRODUCTION", "TRANS_FORMULA", "TRANS_BANK", "TRANS_CASH"]
        - name: direction
          in: query
          schema:
            type: string
            enum: ["DIRECTION_OUT", "DIRECTION_IN", "DIRECTION_TRANSFER"]
        - name: trans_date
          in: query
          schema:
            type: string
            format: date
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - trans
      summary: Create a new trans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/TransPost"
      responses:
        "201":
          description: Trans created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trans"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MetaTags:
      type: array
      items:
        type: string
      description: Additional tags for the data
      example: ["TAG1", "TAG2"]
    DataMap:
      type: object
      additionalProperties: true
      description: Flexible key-value map for additional metadata. The value is any json type.
      example:
        field_name: "Field value"
...
    Rate:
      type: object
      properties:
        rate_type:
          type: string
          enum: [RATE_RATE, RATE_BUY, RATE_SELL, RATE_AVERAGE]
          description: ENUM field specifying the type of rate
        rate_date:
          type: string
          format: date
          description: Date of the rate
        place_code:
          type: string
          description: Reference to Place.code
        currency_code:
          type: string
          description: Reference to Currency.code
        rate_meta:
          type: object
          properties:
            rate_value:
              type: number
              format: double
            tags:
              tags:
              $ref: "#/components/schemas/MetaTags"
        rate_map:
          $ref: "#/components/schemas/DataMap"
    RatePost:
      type: object
      required:
        - rate_type
        - rate_date
        - place_code
        - currency_code
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "RAT1731101982N123"
        - $ref: "#/components/schemas/Rate"
    RateResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/RatePost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
    Tax:
      type: object
      properties:
        tax_meta:
          type: object
          properties:
            description:
              type: string
              example: "VAT 20%"
            rate_value:
              type: number
              format: double
              example: 0.2
            tags:
              $ref: "#/components/schemas/MetaTags"
        tax_map:
          $ref: "#/components/schemas/DataMap"
    TaxPost:
      type: object
      required:
        - code
      allOf:
        - properties:
            code:
              type: string
              description: Unique tax ID.
              example: "VAT20"
        - $ref: "#/components/schemas/Tax"
    TaxResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/TaxPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
    Tool:
      type: object
      properties:
        description:
          type: string
          description: The description of the tool.
          example: "Company car"
        product_code:
          type: string
          description: Reference to Product.code
        tool_meta:
          type: object
          properties:
            serial_number:
              type: string
              example: "ABC-123"
            notes:
              type: string
            inactive:
              type: boolean
            tags:
              $ref: "#/components/schemas/MetaTags"
        tool_map:
          $ref: "#/components/schemas/DataMap"
    ToolPost:
      type: object
      required:
        - description
        - product_code
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "SER1731101982N123"
        - $ref: "#/components/schemas/Tool"
    ToolResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/ToolPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
    TransHead:
      type: object
      properties:
        trans_type:
          type: string
          enum: [TRANS_INVOICE, TRANS_RECEIPT, TRANS_ORDER, TRANS_OFFER, TRANS_WORKSHEET, TRANS_RENT, TRANS_DELIVERY, TRANS_INVENTORY, TRANS_WAYBILL, TRANS_PRODUCTION, TRANS_FORMULA, TRANS_BANK, TRANS_CASH]
          description: Type of transaction
        direction:
          type: string 
          enum: [DIRECTION_OUT, DIRECTION_IN, DIRECTION_TRANSFER]
          description: Direction of transaction
        trans_date:
          type: string
          format: date
        trans_code:
          type: string
          description: Reference to Trans.code
        customer_code:
          type: string
          description: Reference to Customer.code
        employee_code:
          type: string
          description: Reference to Employee.code
        project_code:
          type: string
          description: Reference to Project.code
        place_code:
          type: string
          description: Reference to Place.code
        currency_code:
          type: string
          description: Reference to Currency.code
        auth_code:
          type: string
          description: Reference to Auth.code
    TransMeta:
      type: object
      properties:
        due_time:
          type: string
          format: date-time
        ref_number:
          type: string
        paid_type:
          type: string
          enum: [PAID_ONLINE, PAID_CARD, PAID_TRANSFER, PAID_CASH, PAID_OTHER]
        tax_free:
          type: boolean
        paid:
          type: boolean
        rate:
          type: number
          format: float
        closed:
          type: boolean
        status:
          type: string
          enum: [STATUS_NORMAL, STATUS_CANCELLATION, STATUS_AMENDMENT]
        trans_state:
          type: string
          enum: [STATE_OK, STATE_NEW, STATE_BACK]
        notes:
          type: string
        internal_notes:
          type: string
        report_notes:
          type: string
        worksheet:
          type: object
          properties:
            distance:
              type: number
              format: float
              description: Distance (km)
            repair:
              type: number
              format: float
              description: Repair time (hour)
            total:
              type: number
              format: float
              description: Total time (hour)
            notes:
              type: string
              description: Worksheet justification
        rent:
          type: object
          properties:
            holiday:
              type: number
              format: float
              description: Holidays (day)
            bad_tool:
              type: number
              format: float
              description: Bad tool / machine (hour)
            other:
              type: number
              format: float
              description: Other non-eligible
            notes:
              type: string
              description: Rent justification
    Trans:
      type: object
      allOf:
        - $ref: "#/components/schemas/TransHead"
        - properties:
            trans_meta:
              type: object
              allOf:
                - $ref: "#/components/schemas/TransMeta"
                - properties:
                    tags:
                      $ref: "#/components/schemas/MetaTags"
            trans_map:
              $ref: "#/components/schemas/DataMap"
    TransPost:
      type: object
      required:
        - trans_type
        - direction
        - trans_date
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "INV1731101982N123"
        - $ref: "#/components/schemas/Trans"
    TransResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "INV1731101982N123"
        - $ref: "#/components/schemas/Trans"
        - properties:
            trans_meta:
              type: object
              allOf:
                - $ref: "#/components/schemas/TransMeta"
                - properties:
                    invoice:
                      type: object
                      properties:
                        company_name:
                          type: string
                        company_address:
                          type: string
                        company_tax_number:
                          type: string
                        customer_name:
                          type: string
                        customer_address:
                          type: string
                        customer_tax_number:
                          type: string
                    tags:
                      $ref: "#/components/schemas/MetaTags"
            trans_map:
              $ref: "#/components/schemas/DataMap"
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
...